submission
- [x] Detected Tampered
- [x] Fixed Tampered from Recovery

following changes
- Added CSS to display tampered record
- Added display below which is tampered or not upon verification
- Added node-json-db to store (please read below explanation)
- Added lodash to verify json object

In verify endpoint, response tampered, local, and remote data. This is for the test purpose. However, in real world situation, data will be fixed in the backend and never expose to user and also will call alert such us email and sms to admin for the purpose of follow up necessary actions.